change integers to floats in first 3 katas

+++ Lesson1/IdealWeight.rb
@@ -1,7 +1,7 @@
 puts "What is your name?"
 name = gets.chomp
 puts "What is your height?"
-height = gets.chomp.to_i
+height = gets.chomp.to_f
 
 ideal_weight = (height - 110) * 1.15
